+++ sympy/assumptions/ask_generated.py
@@ -124,6 +124,7 @@ def get_known_facts_cnf():
         Q.integer | ~Q.odd,
         ~Q.algebraic | ~Q.transcendental,
         ~Q.antihermitian | ~Q.hermitian,
+        ~Q.composite | ~Q.prime,
         ~Q.finite | ~Q.infinite,
         ~Q.imaginary | ~Q.real,
         ~Q.irrational | ~Q.rational,
@@ -136,7 +137,6 @@ def get_known_facts_cnf():
         Q.negative | Q.positive | ~Q.nonzero,
         Q.negative | Q.zero | ~Q.nonpositive,
         Q.positive | Q.zero | ~Q.nonnegative,
-        ~Q.prime | ~Q.composite,
         Q.complex_elements | ~Q.real_elements,
         Q.fullrank | ~Q.invertible,
         Q.invertible | ~Q.positive_definite,
